+++ web/src/helpers/redux.js
@@ -64,7 +64,6 @@ function initDataState() {
 
   return {
     failedRequestType: null,
-    failedRequestZoneId: null,
     hasInitializedGrid: false,
     isLoadingHistories: false,
     isLoadingGrid: false,
