+++ web/src/layout/leftpanel/countrypanel.jsx
@@ -18,6 +18,7 @@ import CountryDisclaimer from '../../components/countrydisclaimer';
 import CountryDataInfo from '../../components/countrydatainfo';
 import LoadingPlaceholder from '../../components/loadingplaceholder';
 import Icon from '../../components/icon';
+import { RetryBanner } from '../../components/retrybanner';
 
 import { dispatchApplication } from '../../store';
 
@@ -193,7 +194,7 @@ const CountryHeader = ({ parentPage, zoneId, data, isMobile, isDataAggregated })
 const CountryPanel = ({ electricityMixMode, isMobile, tableDisplayEmissions, zones }) => {
   const [tooltip, setTooltip] = useState(null);
   const { __ } = useTranslation();
-
+  const { failedRequestType } = useSelector((state) => state.data);
   const isLoadingHistories = useSelector((state) => state.data.isLoadingHistories);
 
   // TODO: isLoadingGrid is holding rendering back too much on countryPanel. This should be avoided.
@@ -268,6 +269,7 @@ const CountryPanel = ({ electricityMixMode, isMobile, tableDisplayEmissions, zon
 
   return (
     <CountryPanelStyled>
+      {failedRequestType === 'zone' && <RetryBanner failedRequestType={failedRequestType} />}
       <div id="country-table-header">
         <CountryHeader
           isDataAggregated={timeAggregate && timeAggregate !== TIME.HOURLY}
